Changed comments and text post width and changed edit text post label functionality
Changed the message display width so that, like the text submission
editor, scales along with the window size.

This affects text submissions, text expandos (on frontpage and user
profile) and comments.

Also, added a anchor to the top of the text submission, so that when a
user clicks the edit label, he is taken to the top of the submission.
This is because, currently, when you have a large text submission, if
you click the edit label, due to the edit field having a small height,
the page content will go up, however the user is left down at the bottom
probably looking at comments and not the edit field.