Minor user profile tabs description update
Removed the "To see X, click the X tab" bit from the user profile tabs
description. Mostly to cleanup the page description and also because the
header is where navigation is actually done and if a user doesn't read
the button that says "Comments" or "Submissions", they probably won't
read the description either. Also, if they managed to get to the user
profile, they probably have figured out how to navigate in WhoaVerse.
Additionaly, there's the "To see comments, click the comments tab." in
the overview page, whilst there is nothing mentioning the submissions
tab and IIRC that is there from when the overview tab only showed the
user's submissions (which back then was a bit confusing for those who
were used to reddit overview, since there was also a submitted tab with
the exact same functionality), but since then the overview tab is an
actual user overview, making each tab's button pretty self-explanatory