Ridiculously minor button text update
Changed the "Submitted" button in the user profile header navigation to
"Submissions", since the submission history tab has submissions written
in the title and the url and not "submitted". Also, if it were to be
Submitted, shouldn't the comments button say Commented?

Side note: I shouldn't be commiting so fast, I keep missing something
and have to do another minor commit. :P