Change displayed score to net likes
Suggested change. When a submission is currently voted on, its previous score is updated to either its net likes if liked or net dislikes if disliked. This is a confusing and unnecessary feature, as a submission that has a net score of 250 (say, +250/-0) will display a score of 1 if disliked rather than 249. Similarly, a controversial submission with a net score of 5 (+50/-45) will read 51 once liked instead of 6.