Added "are you sure" check to delete PMs
the functions toggle and toggleback could potentionally be replaced by
togglePM and togglebackPM since the outcome is the same and it prevents
two differents functions doing the same thing. However I can't access
the comments page (due to #684 ) and can't test if it works, so I
haven't done it yet.